pyproject.toml
[tool.poetry]
name = "equilibriumlife"
version = "0.1.0"
description = ""
authors = ["Bohdan Solianyk <bohdan_solianyk@epam.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aioboto3 = "^13.4.0"
openai = "^1.59.9"
python-dotenv = "^1.0.1"
matplotlib = "^3.10.0"
requests = "^2.32.3"
python-telegram-bot = "^21.10"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
moto = "^5.0.27"
pytest-asyncio = "^0.25.2"
types-requests = "^2.32.0.20241016"
types-python-dateutil = "^2.9.0.20241206"
aws-lambda-powertools = "^3.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
